backend_core/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
boto3==1.34.0
opensearch-py==2.4.2
requests-aws4auth==1.2.3
